move pinecone to infrastructure
