+++ packages/core/src/infrastructure/pinecone.ts
